Create Main_Code.ino
This is the code that I made after calibrating the code to what my buttons home position was and then, I also made use of the ezbutton library to be able to click the right button on the controller and then go to a preset home setting. To make the controls work I had to change the if commands for the claw, left arm, right arm, and then turn so that the x and y values that were in the if commands worked the position of my controller. Then to fix the button issue I had to get rid of the code for right_key and left_key as they were also using pins 7 and 8 and had set that up as the z values, so I commented that out so that I could set it so when I click the right button it would go to my preset home setting.