Add files via upload
This commit I added in the if-else statements for the wheels so now I control the wheels while the left joystick is pressed down and am able to move the wheels. I made it so that the right joystick controls going forwards and backward and the left joystick controls side-to-side movement, but I may change it later and make it so that the right joystick just controls all movement so I can keep the left joystick pressed. So it enters the while loop, but when I move the left joystick it doesn't maintain registered as pressed.